deepfake-detection/requirements.txt
requests
python-dotenv  # For handling environment variables